requirements.txt
discord.py
requests
pyyaml
